Make API command safer
The API command was allowing the token to be leaked to anybody able to
trigger the api command. That command was also using a private methind
of the api module, which it shouldn't.

A new method was written for calling custom API methods. It strips out
the token from the response, and also deletes the redundant 'ok'
property.