Better handling of case when bot not in channel
Presence changes aren't recieved unless the bot is in the channel being
used to play IdleRPG, so it's essential that the bot is invited to the
channel. The bot cannot join the channel on its own, as that method is
not callable by bots. So a real user as to send the invite.

The bot now sends a message requesting an invite if it's not in the RPG
channel when it joins. It also skips the step of checking which users
are active, as it would break the game to count anybody as active in
this case (because presence changes aren't recieved, so the active state
isn't accurate as time goes on).

When the bot is invited into a channel, it sends a message of thanks and
re-checks the status of each user.