Implement leveling
Players now level up, based upon the formula x * ( y ^ current_level ).
The values for x and y are traditionally 600 and 1.16 respectively, but
x is set to 300 for now to speed things up a bit.

A new boolean was also added to the bot, active. This tracks whether the
game is currently active or not. This ensures that the game doesn't go
on when the bot isn't in the channel.